Feat: Move less important chat settings into a collapsible panel.
