+++ web/src/pages/agent/form/tool-form/retrieval-form/index.tsx
@@ -45,7 +45,7 @@ const RetrievalForm = () => {
       <FormWrapper>
         <DescriptionField></DescriptionField>
         <MemoryDatasetForm></MemoryDatasetForm>
-        <Collapse title={<div>{t('flow.advancedSettings')}</div>}>
+        <Collapse defaultOpen title={<div>{t('flow.advancedSettings')}</div>}>
           <FormContainer>
             <SimilaritySliderFormField
               similarityWeightName="keywords_similarity_weight"
