+++ web/src/pages/next-chats/chat/app-settings/chat-basic-settings.tsx
@@ -2,10 +2,7 @@
 
 import { AvatarNameDescription } from '@/components/avatar-name-description';
 import { KnowledgeBaseFormField } from '@/components/knowledge-base-item';
-import { MetadataFilter } from '@/components/metadata-filter';
-import { SwitchFormField } from '@/components/switch-fom-field';
-import { TavilyFormField } from '@/components/tavily-form-field';
-import { TOCEnhanceFormField } from '@/components/toc-enhance-form-field';
+import { LlmSettingFieldItems } from '@/components/llm-setting-items/next';
 import {
   FormControl,
   FormField,
@@ -13,84 +10,29 @@ import {
   FormLabel,
   FormMessage,
 } from '@/components/ui/form';
-import { MultiSelect } from '@/components/ui/multi-select';
-import { Switch } from '@/components/ui/switch';
 import { Textarea } from '@/components/ui/textarea';
 import { useTranslate } from '@/hooks/common-hooks';
-import { useFetchKnowledgeMetadataKeys } from '@/hooks/use-knowledge-request';
 import { getDirAttribute } from '@/utils/text-direction';
-import { useEffect, useMemo } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
 
 export default function ChatBasicSetting() {
   const { t } = useTranslate('chat');
   const form = useFormContext();
-  const emptyResponseValue = form.watch('prompt_config.empty_response');
-  const prologueValue = form.watch('prompt_config.prologue');
-  const rawDatasetIds = useWatch({
-    control: form.control,
-    name: 'dataset_ids',
-  });
-  const kbIds = useMemo(
-    () => (rawDatasetIds || []) as string[],
-    [rawDatasetIds],
-  );
-  const metadataInclude = useWatch({
-    control: form.control,
-    name: 'prompt_config.reference_metadata.include',
-  });
-  const { data: metadataKeys, loading: metadataKeysLoading } =
-    useFetchKnowledgeMetadataKeys(kbIds);
-  const metadataFieldOptions = useMemo(() => {
-    return (metadataKeys || []).map((key) => ({
-      label: key,
-      value: key,
-    }));
-  }, [metadataKeys]);
 
-  useEffect(() => {
-    const currentFields = form.getValues(
-      'prompt_config.reference_metadata.fields',
-    );
-    if (
-      metadataInclude &&
-      Array.isArray(currentFields) &&
-      currentFields.length > 0 &&
-      metadataKeys
-    ) {
-      const validFields = currentFields.filter((field) =>
-        metadataKeys.includes(field),
-      );
-      if (validFields.length !== currentFields.length) {
-        form.setValue('prompt_config.reference_metadata.fields', validFields);
-      }
-    } else if (!metadataInclude) {
-      form.setValue('prompt_config.reference_metadata.fields', undefined);
-    }
-  }, [kbIds, metadataKeys, metadataKeysLoading, metadataInclude, form]);
+  const prologueValue = useWatch({
+    control: form.control,
+    name: 'prompt_config.prologue',
+  });
 
   return (
     <div className="space-y-8">
       <AvatarNameDescription />
-      <FormField
-        control={form.control}
-        name={'prompt_config.empty_response'}
-        render={({ field }) => (
-          <FormItem>
-            <FormLabel tooltip={t('emptyResponseTip')}>
-              {t('emptyResponse')}
-            </FormLabel>
-            <FormControl>
-              <Textarea
-                {...field}
-                placeholder={t('emptyResponsePlaceholder')}
-                dir={getDirAttribute(emptyResponseValue || '')}
-              ></Textarea>
-            </FormControl>
-            <FormMessage />
-          </FormItem>
-        )}
-      />
+      <LlmSettingFieldItems
+        prefix="llm_setting"
+        llmId="llm_id"
+        showCollapse
+      ></LlmSettingFieldItems>
+
       <FormField
         control={form.control}
         name={'prompt_config.prologue'}
@@ -109,78 +51,8 @@ export default function ChatBasicSetting() {
           </FormItem>
         )}
       />
-      <SwitchFormField
-        name={'prompt_config.quote'}
-        label={t('quote')}
-        tooltip={t('quoteTip')}
-      ></SwitchFormField>
-      <SwitchFormField
-        name={'prompt_config.keyword'}
-        label={t('keyword')}
-        tooltip={t('keywordTip')}
-      ></SwitchFormField>
-      <SwitchFormField
-        name={'prompt_config.tts'}
-        label={t('tts')}
-        tooltip={t('ttsTip')}
-      ></SwitchFormField>
-      <TOCEnhanceFormField name="prompt_config.toc_enhance"></TOCEnhanceFormField>
-      <TavilyFormField></TavilyFormField>
+
       <KnowledgeBaseFormField></KnowledgeBaseFormField>
-      <MetadataFilter></MetadataFilter>
-      <FormField
-        control={form.control}
-        name={'prompt_config.reference_metadata.include'}
-        render={({ field }) => (
-          <FormItem className="flex flex-row items-start space-x-3 space-y-0">
-            <FormControl>
-              <Switch
-                checked={field.value}
-                onCheckedChange={(value) => {
-                  field.onChange(value);
-                  if (!value) {
-                    form.setValue(
-                      'prompt_config.reference_metadata.fields',
-                      undefined,
-                    );
-                  }
-                }}
-              />
-            </FormControl>
-            <FormLabel tooltip="Display document metadata (e.g., title, page number, upload date) alongside retrieved text chunks">
-              Show chunk metadata
-            </FormLabel>
-          </FormItem>
-        )}
-      />
-      {metadataInclude && (
-        <FormField
-          control={form.control}
-          name={'prompt_config.reference_metadata.fields'}
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel tooltip="Select which metadata fields to display with each chunk">
-                {t('metadataKeys')}
-              </FormLabel>
-              <FormControl className="bg-bg-input">
-                <MultiSelect
-                  options={metadataFieldOptions}
-                  onValueChange={field.onChange}
-                  showSelectAll={false}
-                  placeholder="Please select"
-                  maxCount={20}
-                  defaultValue={Array.isArray(field.value) ? field.value : []}
-                  value={Array.isArray(field.value) ? field.value : []}
-                  name={field.name}
-                  ref={field.ref}
-                  onBlur={field.onBlur}
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-      )}
     </div>
   );
 }
