--- web/src/pages/next-chats/chat/app-settings/chat-model-settings.tsx
@@ -1,12 +1,0 @@
-import { LlmSettingFieldItems } from '@/components/llm-setting-items/next';
-
-export function ChatModelSettings() {
-  return (
-    <div className="space-y-8">
-      <LlmSettingFieldItems
-        prefix="llm_setting"
-        llmId="llm_id"
-      ></LlmSettingFieldItems>
-    </div>
-  );
-}
