+++ web/src/pages/next-chats/chat/app-settings/chat-settings.tsx
@@ -1,7 +1,6 @@
 import { Button } from '@/components/ui/button';
 import { Form } from '@/components/ui/form';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import { Separator } from '@/components/ui/separator';
 import { DatasetMetadata } from '@/constants/chat';
 import { useSetModalState } from '@/hooks/common-hooks';
 import { useFetchChat, useUpdateChat } from '@/hooks/use-chat-request';
@@ -20,7 +19,6 @@ import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router';
 import { z } from 'zod';
 import ChatBasicSetting from './chat-basic-settings';
-import { ChatModelSettings } from './chat-model-settings';
 import { ChatPromptEngine } from './chat-prompt-engine';
 import { SavingButton } from './saving-button';
 import { useChatSettingSchema } from './use-chat-setting-schema';
@@ -195,10 +193,7 @@ export function ChatSettings({ hasSingleChatBox }: ChatSettingsProps) {
                 <ScrollArea viewportClassName="[&>div]:!block">
                   <section className="p-5 space-y-6 overflow-auto flex-1 min-h-0">
                     <ChatBasicSetting></ChatBasicSetting>
-                    <Separator />
                     <ChatPromptEngine></ChatPromptEngine>
-                    <Separator />
-                    <ChatModelSettings></ChatModelSettings>
                   </section>
                 </ScrollArea>
